+++ PhotoDeckAPIXSLT.lua
@@ -28,16 +28,15 @@ return t
 
 PhotoDeckAPIXSLT.websites = xsltheader .. [[
   <xsl:template match='/reply/websites'>
+local t = {}
     <xsl:for-each select='website'>
-local t = { <xsl:value-of select='urlname'/> =
-  {
+t["<xsl:value-of select='urlname'/>"] = {
      hostname = "<xsl:value-of select='hostname'/>",
      homeurl = "<xsl:value-of select='home-url'/>",
      title = "<xsl:value-of select='title'/>",
-  },
 }
-return t
     </xsl:for-each>
+return t
   </xsl:template>
 ]] .. xsltfooter
 
@@ -60,6 +59,27 @@ t["<xsl:value-of select='uuid'/>"] = {
      parentuuid = "<xsl:value-of select='parent-uuid'/>",
 }
    </xsl:for-each>
+return t
+  </xsl:template>
+]] .. xsltfooter
+
+PhotoDeckAPIXSLT.gallery = xsltheader .. [[
+  <xsl:template match='/reply/gallery'>
+local t = {
+     fullurlpath = "<xsl:value-of select='full-url-path'/>",
+     name = "<xsl:value-of select='name'/>",
+     uuid = "<xsl:value-of select='uuid'/>",
+     urlpath = "<xsl:value-of select='url-path'/>",
+}
+return t
+  </xsl:template>
+]] .. xsltfooter
+
+PhotoDeckAPIXSLT.createGallery = xsltheader .. [[
+  <xsl:template match='/reply'>
+local t = {
+     uuid = "<xsl:value-of select='gallery-uuid'/>",
+}
 return t
   </xsl:template>
 ]] .. xsltfooter